\name{credicpp-package}
\alias{credicpp-package}
\alias{credicpp}
\docType{package}
\title{
\packageTitle{credicpp}
}
\description{
\packageDescription{credicpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{credicpp}
\packageIndices{credicpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{credicpp}

Maintainer: \packageMaintainer{credicpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
